Update Create-UCSTechSupport.ps1
Hi David,
I was verry lucky to find your code last week as I always find it annoying to download all the bundles for the ucs manager and also for all the chassis one by one if I have to open a service request at cisco for lets say the next firmware update. I added some small things which are making it for me very useful and i wanted to share this with you. Maybe it helps you too, if it does not work or you find it annoying feel free to ignore my input :-). The Idea was that i can start the script and it downloads me all the chassis files and also the ucsm file to a specific folder on the local machine together with a text file which contains the information of the UCS Domain (FI- Serial number) which i usually use to create a tech support file. So that i can run the script and afterwards have all the tech support files and also the FI Serial at hand to open a service request. It worked in our environment so i think it is adaptable. Maybe you can come up with a better idea regarding the "overwriting the files if they already exist" method as i did not spent a lot of effort there.

Let me know what u think.

All the Best
Stephan